requirements.txt
python-telegram-bot==11.1.0
flask
lxml==4.6.2
pychromecast==9.0.0
tqdm==4.29.1